Refactor Contentful rendering and Graphql queries
- Extracted common Graphql fragments for Contentful components, enhancing maintainability and reducing redundancy.
- Abstracted rendering logic for rich text fields into a utility module, ensuring clearer separation of concerns and making the rendering process more modular.
- Updated Contentful image queries to use 'CONSTRAINED' layout with 'BLURRED' placeholders for better loading performance and user experience.
- Created utility functions to handle tab generation logic, simplifying embedded content handling.
- Removed unused imports and simplified ContentfulPageTemplate, improving code readability and component efficiency.

This refactor makes future updates to Contentful content types and embedded entities easier to manage, while also improving code clarity and front-end performance.

#123 Refactor Contentful Components
