+++ src/utils/index.tsx
@@ -1,0 +1,2 @@
+export * from './getTabs';
+export * from './options';
